health-check/requirements.txt
connexion[flask]
httpx
pyyaml
starlette
uvicorn